refactor: rename types from 'assets/mock'
