+++ src/assets/mock/pageMockNewsList.ts
